ai-content-engine/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
Jinja2==3.1.2
SQLAlchemy==1.4.49
alembic==1.11.1
celery==5.3.1
redis==4.5.3
gunicorn==20.1.0
requests==2.31.0
feedparser==6.0.10
slowapi==0.1.9
python-jose[cryptography]==3.3.0
pytrends==4.9.2
